Rename SharesDatabase to PoolDatabase
